+++ example/src/App.tsx
@@ -39,6 +39,9 @@ import { AmplitudeSessionPlugin } from '@segment/analytics-react-native-plugin-a
 //@ts-ignore
 // import { AdvertisingIdPlugin } from '@segment/analytics-react-native-plugin-advertising-id';
 
+//@ts-ignore
+// import { ClevertapPlugin } from '@segment/analytics-react-native-plugin-clevertap';
+
 const segmentClient = createClient({
   writeKey: 'WRITE_KEY',
   trackAppLifecycleEvents: true,
@@ -69,6 +72,8 @@ segmentClient.add({ plugin: LoggerPlugin });
 // segmentClient.add({ plugin: idfaPlugin });
 
 segmentClient.add({ plugin: new AmplitudeSessionPlugin() });
+
+// segmentClient.add({ plugin: new ClevertapPlugin() });
 
 // segmentClient.add({
 //   plugin: new AdvertisingIdPlugin(),