+++ packages/core/src/__tests__/__helpers__/setupSegmentClient.ts
@@ -1,6 +1,6 @@
 import { SegmentClient } from '../../analytics';
 import { UtilityPlugin } from '../../plugin';
-import { PluginType, SegmentEvent } from '../../types';
+import { Config, PluginType, SegmentEvent } from '../../types';
 import { getMockLogger } from './mockLogger';
 import { MockSegmentStore, StoreData } from './mockSegmentStore';
 
@@ -10,7 +10,10 @@ jest
   .spyOn(Date.prototype, 'toISOString')
   .mockReturnValue('2010-01-01T00:00:00.000Z');
 
-export const createTestClient = (storeData?: Partial<StoreData>) => {
+export const createTestClient = (
+  storeData?: Partial<StoreData>,
+  config?: Partial<Config>
+) => {
   const store = new MockSegmentStore({
     isReady: true,
     ...storeData,
@@ -20,15 +23,22 @@ export const createTestClient = (storeData?: Partial<StoreData>) => {
     config: {
       writeKey: 'mock-write-key',
       autoAddSegmentDestination: false,
+      ...config,
     },
     logger: getMockLogger(),
     store: store,
   };
 
   const client = new SegmentClient(clientArgs);
-
   class ObservablePlugin extends UtilityPlugin {
     type = PluginType.after;
+
+    override execute(
+      event: SegmentEvent
+    ): SegmentEvent | Promise<SegmentEvent | undefined> | undefined {
+      super.execute(event);
+      return event;
+    }
   }
 
   const mockPlugin = new ObservablePlugin();